+++ resources/js/Pages/Campaigns/Campaign.vue
@@ -1,5 +1,6 @@
 <script setup>
 import Dropdown from '@/Components/Dropdown.vue';
+import NavLink from '@/Components/NavLink.vue';
 import PrimaryButton from '@/Components/PrimaryButton.vue';
 import AppLayout from '@/Layouts/AppLayout.vue';
 import { useForm } from '@inertiajs/vue3';
@@ -38,13 +39,18 @@ const importContact = () => {
 <template>
     <AppLayout title="Email">
         <main class="min-h-screen">
-            <div class="mx-auto mt-5 rounded-md max-w-4xl px-4 py-6 sm:px-6 lg:px-8">
-                <form @submit.prevent="importContact">
+            <div class="mx-auto mt-5 rounded-md max-w-7xl px-4 py-6 sm:px-6 lg:px-8">
+                <!-- <form @submit.prevent="importContact">
                     <h1 class="text-lg font-semibold">Import File</h1>
                     <input type="file" @change="form.file = $event.target.files[0]">
                     <button type="submit">Import</button>
-                </form>
-
+                </form> -->
+                <div class="">
+                    <NavLink :href="route()"
+                    class="bg-primary text-white font-medium text-sm px-6 py-2 rounded-md hover:bg-white hover:text-primary hover:border-primary transition mr-4">
+                        Create Campaign
+                    </NavLink>
+                </div>
                 <form class="bg-white p-3 rounded-md" @submit.prevent="submit">
                     <h1 class="text-lg font-semibold">Email Send</h1>
                     <input type="email" name="" id=""
